backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
celery[redis]==5.3.6
redis==5.0.4
anthropic==0.26.0
ultralytics==8.2.2
opencv-python-headless==4.9.0.80
ffmpeg-python==0.2.0
boto3==1.34.95
numpy==1.26.4
pytest==8.2.0
pytest-asyncio==0.23.6
